+++ packages/ariakit-test-utils/src/e2e.ts
@@ -1,8 +1,0 @@
-import { basename } from "path";
-
-export function navigateToExamplePage(dirname: string) {
-  const exampleName = basename(dirname);
-  const baseUrl = process.env.BASE_URL || "http://localhost:3000";
-  const url = `${baseUrl}/examples/${exampleName}`;
-  beforeAll(() => page.goto(url));
-}