+++ packages/ariakit/src/checkbox/__examples__/checkbox-state/test.tsx
@@ -1,10 +1,4 @@
-import {
-  click,
-  getByLabelText,
-  getByRole,
-  press,
-  render,
-} from "ariakit-test-utils";
+import { click, getByLabelText, getByRole, press, render } from "ariakit-test";
 import { axe } from "jest-axe";
 import Example from ".";