+++ packages/ariakit/src/menu/__examples__/menu-bar/test.tsx
@@ -7,7 +7,7 @@ import {
   render,
   sleep,
   type,
-} from "ariakit-test-utils";
+} from "ariakit-test";
 import Example from ".";
 
 const getMenu = (name: string) =>