+++ packages/ariakit/src/menu/__examples__/menu-nested/test.tsx
@@ -8,7 +8,7 @@ import {
   sleep,
   type,
   waitFor,
-} from "ariakit-test-utils";
+} from "ariakit-test";
 import Example from ".";
 
 const getMenuButton = () => getByRole("button", { name: "Edit" });