+++ packages/ariakit/src/toolbar/__examples__/toolbar/test.tsx
@@ -1,4 +1,4 @@
-import { getByRole, press, render } from "ariakit-test-utils";
+import { getByRole, press, render } from "ariakit-test";
 import { axe } from "jest-axe";
 import Example from ".";