+++ packages/ariakit/src/tooltip/__examples__/tooltip-timeout/test.tsx
@@ -1,11 +1,4 @@
-import {
-  blur,
-  getByRole,
-  hover,
-  press,
-  render,
-  waitFor,
-} from "ariakit-test-utils";
+import { blur, getByRole, hover, press, render, waitFor } from "ariakit-test";
 import Example from ".";
 
 test("show tooltip on hover after timeout", async () => {