+++ packages/website/pages/_document.tsx
@@ -6,6 +6,24 @@ import Document, {
   NextScript,
 } from "next/document";
 
+const fontFaces = `
+@font-face {
+  font-family: Inter;
+  font-weight: 100 900;
+  font-display: swap;
+  font-style: normal;
+  font-named-instance: 'Regular';
+  src: url("/fonts/Inter-roman.var.woff2?v=3.19") format("woff2");
+}
+@font-face {
+  font-family: Inter;
+  font-weight: 100 900;
+  font-display: swap;
+  font-style: italic;
+  font-named-instance: 'Italic';
+  src: url("/fonts/Inter-italic.var.woff2?v=3.19") format("woff2");
+}`;
+
 const darkModeScript = `
 function classList(action) {
   document.documentElement.classList[action]("dark");
@@ -28,7 +46,16 @@ class MyDocument extends Document {
   render() {
     return (
       <Html>
-        <Head />
+        <Head>
+          <link
+            rel="preload"
+            crossOrigin="anonymous"
+            href="/fonts/Inter-roman.var.woff2?v=3.19"
+            type="font/woff2"
+            as="font"
+          />
+          <style dangerouslySetInnerHTML={{ __html: fontFaces }} />
+        </Head>
         <body>
           <script dangerouslySetInnerHTML={{ __html: darkModeScript }} />
           <Main />